// interfaces/product.interface.ts
export interface Product {
  id: string;
  handle: string;
  type: string;
  stock: number;
  title: string;
  description: string;
  hightlights: string;
  details: string;
  reviewsNumber: number;
  imageUrl1: string;
  imageUrl2: string;
  imageUrl3: string;
  imageUrl4: string;
  minVariantPrice: string;
  maxVariantPrice: string;
}